Refactor squash and approve
